refactor: extract into `CleanupTaxon`
Because `Taxon` have become too much of a dumping ground even for my
taste :)

It's still there, just slightly more separated :)
