+++ app/views/taxa/_form/_main_form.haml
@@ -55,7 +55,7 @@
 
   %tr
     %td=form.label :origin
-    %td=form.select :origin, [['(none)', nil], *Taxon::ORIGINS]
+    %td=form.select :origin, [['(none)', nil], *CleanupTaxon::ORIGINS]
 
   %tr
     %td
